# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/stream/upload": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Upload to a remote a single file",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "example": "any"
                },
                "fileName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result of the operation",
            "schema": {
              "type": "object",
              "properties": {
                "files": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "xxx.mp4"
                      },
                      "size": {
                        "type": "number",
                        "example": 11111
                      },
                      "url": {
                        "type": "string",
                        "example": "https://domain.com/mcurg8n382uy"
                      },
                      "deleteUrl": {
                        "type": "string",
                        "example": "https://domain.com/mcurg8n382uy?killcode=agdt0meepz"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Method Not Allowed"
          }
        }
      }
    },
    "/stream/check-existing": {
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Upload to a remote a single file",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "magnet": {
                  "example": "any"
                },
                "fileName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result of the operation",
            "schema": {
              "type": "object",
              "properties": {
                "files": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "example": "xxx.mp4"
                      },
                      "size": {
                        "type": "number",
                        "example": 11111
                      },
                      "url": {
                        "type": "string",
                        "example": "https://domain.com/mcurg8n382uy"
                      },
                      "deleteUrl": {
                        "type": "string",
                        "example": "https://domain.com/mcurg8n382uy?killcode=agdt0meepz"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "405": {
            "description": "Method Not Allowed"
          }
        }
      }
    },
    "/stream/config": {
      "get": {
        "tags": [
          "Stream"
        ],
        "summary": "Retrieve the configuration about streaming",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuration data",
            "schema": {
              "type": "object",
              "properties": {
                "uptobox": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "xxxxxxx"
                    }
                  }
                }
              },
              "required": [
                "uptobox"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stream"
        ],
        "summary": "Update the configuration about streaming platform",
        "description": "",
        "parameters": [
          {
            "name": "conf",
            "in": "body",
            "description": "Data about configuration (Only uptobox available now",
            "schema": {
              "type": "object",
              "properties": {
                "uploadEnabled": {
                  "type": "boolean",
                  "example": false
                },
                "uptobox": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "xxxxxxx"
                    }
                  }
                }
              },
              "required": [
                "uploadEnabled",
                "uptobox"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration data",
            "schema": {
              "type": "object",
              "properties": {
                "uptobox": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "xxxxxxx"
                    }
                  }
                }
              },
              "required": [
                "uptobox"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/file/stream/{filename}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Open the file in the local system",
        "description": "",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fileid",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}